feat: Add usekeywords. Cleanpu useSEO
