+++ ui/routes/Home/index.tsx
@@ -1,11 +1,12 @@
 import { RouteComponentProps } from '@reach/router'
 import React from 'react'
 import { useTitle } from 'ui/components/HeadProvider'
-import { useKeywords } from 'ui/components/useSEO'
+import { useKeywords, useDescription } from 'ui/components/useSEO'
 
 export function HomeRoute(props: RouteComponentProps) {
   useTitle('Index Page')
   useKeywords(['Defrex', 'KristianFJones'])
+  useDescription('Firebase React SSR Demo')
   return (
     <div>
       <h1>Index</h1>
